Improve code 2

+++ app/src/main/java/com/neoterm/setup/SetupThread.java
@@ -2,7 +2,7 @@ package com.neoterm.setup;
 
 import android.app.ProgressDialog;
 import android.system.Os;
-import android.util.Pair;
+import androidx.core.util.Pair;
 import androidx.appcompat.app.AppCompatActivity;
 import com.termux.terminal.EmulatorDebug;
 import com.neoterm.component.config.NeoTermPath;
@@ -56,7 +56,7 @@ final class SetupThread extends Thread {
         int totalBytes = sourceConnection.getSize();
 
         while ((zipEntry = zipInput.getNextEntry()) != null) {
-          totalReadBytes += zipEntry.getCompressedSize();
+          totalReadBytes += (int) zipEntry.getCompressedSize();
 
           final int totalReadBytesFinal = totalReadBytes;
           final int totalBytesFinal = totalBytes;
